requirements.txt
torch
torchvision
torchaudio
gTTS
basicsr==1.4.2
gfpgan==1.3.8
pydub
flask
moviepy==1.0.3
imageio==2.31.1
imageio-ffmpeg==0.4.8
google-generativeai
